requirements.txt
python-telegram-bot==20.3
pandas==2.1.0
openpyxl==3.1.2
